requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15.2,<3.16
psycopg2>=2.9.9,<2.10
drf-spectacular>=0.27.2,<0.28
Pillow>=10.4.0,<10.5